+++ components/Cards/vehicle-armor-card.tsx
